+++ workflow/scripts/run_polegon.py
@@ -10,10 +10,12 @@ import yaml
 import subprocess
 import numpy as np
 import msprime
-import tszip
 import tskit
 import pickle
 from datetime import datetime
+
+from utils import clip_and_shift_intervals
+from utils import effective_edge_spans
 
 # --- lib --- #
 
@@ -21,25 +23,134 @@ def tag():
     return f"[singer-snakemake::{snakemake.rule}::{str(datetime.now())}]"
 
 
+def singer_to_tree_sequence(
+    node_file: str, 
+    branch_file: str, 
+    mutation_file: str,
+) -> tskit.TreeSequence:
+    # build nodes and edges tables
+    node_time = np.loadtxt(node_file)
+    branches = np.loadtxt(branch_file)
+    left, right, parent, child = branches.T
+    parent = parent.astype(np.int32)
+    child = child.astype(np.int32)
+    tables = tskit.TableCollection(sequence_length=right.max())
+    for time in node_time:
+        if time == 0:
+            tables.nodes.add_row(flags=tskit.NODE_IS_SAMPLE)
+        else:
+            tables.nodes.add_row(time=time)
+    valid = parent >= 0
+    tables.edges.set_columns(
+        left=left[valid],
+        right=right[valid], 
+        parent=parent[valid],
+        child=child[valid],
+    )
+    # build mutations and sites tables
+    mutations = np.loadtxt(mutation_file)
+    last_position = -1
+    last_site = 0
+    for (pos, node, _, state) in mutations:
+        if pos != last_position and pos < tables.sequence_length:
+            last_site = tables.sites.add_row(position=pos, ancestral_state='0')
+            last_position = pos 
+        tables.mutations.add_row(
+            site=last_site, 
+            node=int(node), 
+            derived_state=str(int(state)),
+        )
+    # sort and finalize
+    tables.sort()
+    tables.build_index()
+    tables.compute_mutation_parents()
+    tables.compute_mutation_times()
+    return tables.tree_sequence()
+
+
+def root_table(ts: tskit.TreeSequence) -> np.ndarray:
+    # FIXME: different logic is needed to handle gaps
+    left, root = 0.0, tskit.NULL
+    stems = []
+    for tree in ts.trees():
+        assert tree.num_edges
+        right = tree.interval.left
+        next_root = tree.root if tree.num_edges else tskit.NULL
+        if next_root != root:
+            stems.append([left, right, tskit.NULL, root])
+            left, root = right, next_root
+    stems.append([left, ts.sequence_length, tskit.NULL, root])
+    stems = np.array(stems).reshape(-1, 4)
+    return stems
+
+
+def mutation_table(ts: tskit.TreeSequence) -> np.ndarray:
+    edges = ts.mutations_edge
+    positions = ts.sites_position[ts.mutations_site]
+    children = ts.edges_child[edges]
+    parents = ts.edges_parent[edges]
+    derived_states = np.array([
+        int(s) for s in tskit.unpack_strings(
+            ts.tables.mutations.derived_state,
+            ts.tables.mutations.derived_state_offset,
+        )
+    ])
+    mut_table = np.stack([positions, children, parents, derived_states], axis=-1)
+    return mut_table
+
+
+
 # --- implm --- #
 
 logfile = open(snakemake.log.log, "w")
 use_polegon = snakemake.params.use_polegon
 use_mutational_span = snakemake.params.use_mutational_span
+remove_polegon_inputs = snakemake.params.remove_polegon_inputs
 drop_omitted = snakemake.params.drop_omitted
+skip_node_masks = snakemake.params.skip_node_masks
+node_masks = pickle.load(open(snakemake.input.node_masks, "rb"))
+mut_rate = pickle.load(open(snakemake.input.mut_rate, "rb"))
 failed_chunk = os.path.getsize(snakemake.input.nodes) == 0
 
 if use_polegon and not failed_chunk:
-    singer_params = yaml.safe_load(open(snakemake.input.params)).pop("singer")
-    polegon_params = yaml.safe_load(open(snakemake.input.params)).pop("polegon")
+    params_yaml = yaml.safe_load(open(snakemake.input.params))
+    singer_params = params_yaml.pop("singer")
+    polegon_params = params_yaml.pop("polegon")
     seed = polegon_params.pop("seed") + int(snakemake.wildcards.rep)
-
-    # FIXME: Ne is lower by factor of two relative to `polegon_master`.
-    # This shouldn't matter, it cancels during rescaling, but look into it.
-
-    # POLEGON expects inputs named slightly differently than SINGER output
+    start, end = singer_params["start"], singer_params["end"]
     prefix = snakemake.input.muts.replace("_muts_", "_").removesuffix(".txt")
-    shutil.copy(snakemake.input.nodes, f"{prefix}_nodes.txt")
+
+    treeseq = singer_to_tree_sequence(
+        snakemake.input.nodes, 
+        snakemake.input.branches, 
+        snakemake.input.muts,
+    )
+    logfile.write(f"{tag()} Read tree sequence from SINGER input:\n{treeseq}\n")
+
+    # Absorb mutation rate into span.  This is necessary because POLEGON
+    # doesn't use the "average branch mutation rate" from the mutation map
+    # during rescaling, and instead just adjusts by the mean rate. Manually
+    # converting the branch spans to mutational units fixes this.
+    if use_mutational_span:
+        logfile.write(
+            f"{tag()} Adjusting coordinate system to reflect mutation rate map "
+            f"and setting mutation rate to unity\n"
+        )
+        polegon_params["m"] = 1.0
+        adjusted_mu = mut_rate.slice(start, end, trim=True)
+    else:
+        adjusted_mu = msprime.RateMap(position=[0, treeseq.sequence_length], rate=[1.0])
+
+    node_masks = None if skip_node_masks else {
+        sample: clip_and_shift_intervals(intervals, start, end)
+        for sample, intervals in node_masks.items()
+    }
+    assert np.all(adjusted_mu.get_rate(treeseq.sites_position) > 0)
+    retained_edge_span, retained_mutations = effective_edge_spans(treeseq, adjusted_mu, node_masks)
+    logfile.write(
+        f"{tag()} Retained {retained_mutations.sum()} of {retained_mutations.size} "
+        f"mutations after sample masking\n"
+    )
 
     # Adjust mutation list to omit specified sites. These will not be used for
     # dating, and are effectively inaccessible sequence. However, they were used to
@@ -49,46 +160,36 @@ if use_polegon and not failed_chunk:
         # should be subtracted from edge spans-- but if we assume omitted sites
         # are a very small proportion of edge spans, then the bias should be
         # minimal.
-        left, right = singer_params["start"], singer_params["end"]
         omitted_positions = pickle.load(open(snakemake.input.omitted, "rb"))
-        in_bounds = np.logical_and(omitted_positions >= left, omitted_positions < right)
-        omitted_positions = omitted_positions[in_bounds] - left
-        logfile.write(f"{tag()} Read list of {omitted_positions.size} omitted positions\n")
-        mutations = np.loadtxt(snakemake.input.muts)
-        omitted_mutations = np.isin(mutations[:, 0].astype(np.int64), omitted_positions)
+        in_bounds = np.logical_and(omitted_positions >= start, omitted_positions < end)
+        omitted_positions = omitted_positions[in_bounds] - start
+        omitted_mutations = np.isin(treeseq.sites_position[treeseq.mutations_site], omitted_positions)
         logfile.write(f"{tag()} Omitting {omitted_mutations.sum()} mutations from dating\n")
-        np.savetxt(f"{prefix}_muts.txt", mutations[~omitted_mutations])
-    else:
-        shutil.copy(snakemake.input.muts, f"{prefix}_muts.txt")
+        retained_mutations = np.logical_and(retained_mutations, ~omitted_mutations)
 
-    # Adjust branch spans to reflect masked sequence, and absorb mutation rate
-    # into span.  This is necessary because POLEGON doesn't use the "average
-    # branch mutation rate" from the mutation map during rescaling, and instead
-    # just adjusts by the mean rate. Manually converting the branch spans to
-    # mutational units fixes this.
-    if use_mutational_span:
-        logfile.write(
-            f"{tag()} Adjusting branch spans input ({prefix}_branches.txt) "
-            f"to reflect mutation rate map and setting mutation rate to unity\n"
-        )
-        polegon_params["m"] = 1.0
-        mutation_map = polegon_params.pop("mutation_map")
-        adjusted_mu = np.loadtxt(mutation_map, ndmin=2)
-        assert np.all(adjusted_mu[:-1, 1] == adjusted_mu[1:, 0])
-        adjusted_mu = msprime.RateMap(
-           position=np.append(adjusted_mu[0, 0], adjusted_mu[:, 1]),
-           rate=adjusted_mu[:, 2],
-        )
-        branches = np.loadtxt(snakemake.input.branches)
-        for i in range(2):
-           branches[:, i] = adjusted_mu.get_cumulative_mass(branches[:, i])
-        np.savetxt(f"{prefix}_branches.txt", branches)
-    else:
-        logfile.write(
-            f"{tag()} Using mutation rate map without adjusting branch spans input "
-            f"(for debugging only, will likely produce biased node ages)\n"
-        )
-        shutil.copy(snakemake.input.branches, f"{prefix}_branches.txt")
+    # Assemble POLEGON inputs. POLEGON only cares about edge boundaries insofar
+    # as these impact edge span.
+    retained_mutations = np.logical_and(retained_mutations, treeseq.mutations_edge != tskit.NULL)
+    edge_left = adjusted_mu.get_cumulative_mass(treeseq.edges_left)
+    edge_right = edge_left + retained_edge_span
+    edge_table = np.column_stack([edge_left, edge_right, treeseq.edges_parent, treeseq.edges_child])
+    root_stems = root_table(treeseq)
+    root_stems[:, :2] = adjusted_mu.get_cumulative_mass(root_stems[:, :2])
+    root_stems = root_stems[root_stems[..., -1] != tskit.NULL]
+    branches = np.concatenate([edge_table, root_stems], axis=0)
+    mutations = mutation_table(treeseq)[retained_mutations]
+    nodes = treeseq.nodes_time
+    logfile.write(
+        f"{tag()} POLEGON input dimensions: {nodes.size} nodes, {mutations.shape[0]} mutations, "
+        f"{edge_table.shape[0]} edges, {root_stems.shape[0]} roots\n"
+    )
+
+    polegon_nodes = f"{prefix}_nodes.txt"
+    polegon_branches = f"{prefix}_branches.txt"
+    polegon_muts = f"{prefix}_muts.txt"
+    np.savetxt(polegon_nodes, nodes)
+    np.savetxt(polegon_branches, branches)
+    np.savetxt(polegon_muts, mutations)
 
     invocation = [
         f"{snakemake.params.polegon_binary}",
@@ -102,8 +203,8 @@ if use_polegon and not failed_chunk:
     process = subprocess.run(invocation, check=False, stdout=logfile, stderr=logfile)
     logfile.write(f"{tag()} POLEGON run ended ({process.returncode})\n")
 
-    for suffix in ["muts", "branches", "nodes"]:
-        os.remove(f"{prefix}_{suffix}.txt")
+    for file in [polegon_nodes, polegon_branches, polegon_muts]: 
+        if remove_polegon_inputs: os.remove(file)
 
     assert process.returncode == 0, f"POLEGON terminated with error ({process.returncode})"
     os.rename(f"{prefix}_new_nodes.txt", snakemake.output.nodes)